requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.8.3
charset-normalizer==3.4.2
greenlet==3.2.3
idna==3.10
python-telegram-bot==13.5
pytz==2025.2
requests==2.32.4
six==1.17.0
soupsieve==2.7
SQLAlchemy==2.0.42
tornado==6.5.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
